Develop the ScriptCommitment wrapper type.
The ScriptCommitment type is a container struct that holds the
(secret and public) information necessary to derive the corresponding
Pedersen commitment.

The script commitment is a modified Pedersen commitment of the form

SC = C + H(C||s) where

SC is the script commitment
C is the "naked" pedersen commitment
H is a hash function (with 32 byte digest)
s is the hash of the serialised Tari script

To achieve this:
First, we add the TariScript::as_hash function, which lets us easily
calculate s.

Then we add the `ScriptCommitmentFactory` type that takes a blinding
factor, value and Tari script and calculates the intermediate terms in
the equation above, returning them in a `ScriptCommitment` struct.

The `ScriptCommitmentFactory` can take this object and generate the
equivalent Pedersen commitment with the `script_to_pedersen` method.