Add private key utility functions to the WASM interface
This PR adds a number of methods to the WASM interface to create private keys and do some basic private key arithmetic.  This allows a client to aggregate private keys. This is being used in the cucumber tests in the main Tari repo to build the components required for TariScript transactions.