Migrate to digest 0.9
This PR moves all dependencies to use the digest 0.9 traits and APIs.
This is a breaking change, so the minor version is incremented.

Clients of this generally only need to update the `result` method to
`finalize`; and obviously make use of the v0.9 `digest::Digest` trait
where necessary.

As a result, the deprecated k12, sha3 and Blake3 objects can be removed.
Methods and functins that need a hasher are all generic over `Digest`.

We retain the convenience wrapper over `VarBlake2B` to produce 256 bit
hashes and implement the necessary sub-traits to support
`digest::Digest`.

This update also fixes
https://github.com/tari-project/tari-crypto/issues/35