Source coverage (#48)
* Support stable rust

Updated dependencies such that Rust stable 1.53 is now supported.
The optimised avx_2 option will NOT rust on stable because there's
still an unstable feature on subtle-ng. BUT this feature is actually
for doc generation and has been removed from Rust. As soon as subtle-ng
merges https://github.com/dalek-cryptography/subtle/pull/85, avx2 will
probably be supported on stable as well.

The rand dep update breaks the benchmarks, but these will be fixed in
the next commit.

* Update benchmarks and fmt

* Update benchmark code to use new dependency APIs
* Run rust-fmt

* Use source coverage in tests

Replace the grcov action with the more accurate source coverage approach
as detailed in
https://doc.rust-lang.org/nightly/unstable-book/compiler-flags/instrument-coverage.html

This PR also provides the `test_coverage.sh` script that generates a
test coverage report lcoally (see the comments for setting up
pre-requisites)

* Exclude std lib from coverage