feat: `RistrettoSecretKey` inversion (#173)
Adds multiplicative inversion to `RistrettoSecretKey`. Adds a few unit
tests.

Note that the underlying `invert` functionality from `curve25519-dalek`
does not assert that the `Scalar` in question is nonzero, but requires
the caller to check this; inverting zero is undefined. This PR
specifically checks this, and as a result, the inversion returns an
`Option`. This is probably annoying for the caller, but might stop bad
things from happening.