fix!: improve zeroizing support (#209)
This adds `Zeroize` and `ZeroizeOnDrop` to `ExtendedMask` and
`ExtendedWitness` for improved memory handling functionality. It
restricts the visibility of an internal commitment constructor and adds
zeroizing to temporary secret values. It updates secret key
deserialization to add zeroizing to a temporary byte array. Finally, it
removes a secret key constructor.

Supersedes #204.

BREAKING CHANGE: Changes the commitment and secret key APIs.